+++ portfolio_engine/optimizers/__init__.py
@@ -6,6 +6,7 @@ from portfolio_engine.optimizers.base import (
     SetupInfo,
     UpdateInfo,
 )
+from portfolio_engine.optimizers.feasibility_oracle import FeasibilityVerdict, lp_feasibility
 from portfolio_engine.optimizers.highs_backend import HiGHSLPBackend
 from portfolio_engine.optimizers.osqp_backend import OSQPBackend
 from portfolio_engine.optimizers.problem import CanonicalProblem
@@ -20,12 +21,14 @@ __all__ = (
     "AMBIGUOUS_STATUSES",
     "BackendCapabilities",
     "CanonicalProblem",
+    "FeasibilityVerdict",
     "HiGHSLPBackend",
     "OSQPBackend",
     "OptimizationBackend",
     "SetupInfo",
     "SolverRouter",
     "UpdateInfo",
+    "lp_feasibility",
     "map_highs_status",
     "map_osqp_status",
     "solve_with_policy",